+++ CHANGELOG.rst
@@ -40,6 +40,7 @@ Added
 * Job directive (for ``url`` jobs) ``no_redirects``: disable GET/OPTIONS/POST/PUT/PATCH/DELETE/HEAD redirection
   (true/false). Suggested by `snowman <https://github.com/snowman>`__ upstream `here
   <https://github.com/thp/urlwatch/issues/635>`__.
+* Alert when the job file contains unrecognized directives (e.g. typo)
 
 Fixed
 -----